qemu_v8: clone DTC source
Ubuntu 1.4.04 LTS, 16.04 LTS and even 17.04 provide the libfdt in
version 1.4.0. This version is too old for the latest QEMU that
requires libfdt in version 1.4.2 or higher.

This change updates qemu_v8.xml so that the qemu/dtc/ source code is
cloned locally and therefore the build does not depend on libfdt.

Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Reviewed-by: Igor Opaniuk <igor.opaniuk@linaro.org>