Update TF-A to v2.2 (all platforms but QEMU)
Updates all platforms except QEMU to TF-A to v2.2. This notably
includes support for using CNTPCT in TAs [2], and therefore fixes an
issue with function tracing:

 $ make CFG_FTRACE_SUPPORT=y CFLAGS_ta_arm32="-O0 -pg"
 ...
 # xtest 4001
 ...
 E/TC:? 0 User TA undef-abort at address 0x1921e8 ftrace_enter+284 .text+45512
 E/LD:  Call stack:
 E/LD:   0x001921e8 read_cntpct at optee_os/out/arm/include/generated/arm32_user_sysreg.h:33
 E/LD:   0x00191f10 __gnu_mcount_nc at optee_os/lib/libutils/ext/arch/arm/mcount_a32.S:43

HiKey960: the OpenPlatformPkg repository is updated at the same time or
the board would hang on boot after the TF-A BL31 initialization message:
"NOTICE:  BL31: Built : <time and date>".

QEMU is not updated due to a regression in TF-A fixed in [2].

Link: [1] https://git.trustedfirmware.org/TF-A/trusted-firmware-a.git/commit?id=43f999a7e35d
Link: [2] https://review.trustedfirmware.org/c/TF-A/trusted-firmware-a/+/3388
Change-Id: I20bcfbe1c9f2ec815e4430fabc3c6f9c9cda4168
Signed-off-by: Jerome Forissier <jerome@forissier.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Tested-by: Jerome Forissier <jerome@forissier.org> (HiKey)
Tested-by: Jerome Forissier <jerome@forissier.org> (HiKey960)
Tested-by: Jerome Forissier <jerome@forissier.org> (QEMUv8)