repo: remove symlink creation to gdb
It seems like the newest version of repo (v2.0) no longer allows to
create symlinks to non-existent files. In the past repo has silently
ignored creating the symlink on the first repo sync. However the most
recent version gives an error instead.

Since our setup flow is to create toolchains after checking out the
repo forest, the only viable option we have is to no longer try to
create a symlink. This doesn't really affect anything as such, we just
had it there as a convenience so we could handle builds and debugging
from the same folder on disc.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jerome Forissier <jerome@forissier.org>