make_stable.sh: only modify "project" elements
Currently the script doesn't have filtering based on which type of
element the line contains. This can lead to unexpected results or syntax
errors, e.g. when using a "remove-project" element [1].

Fix this issue by introducing an address [2] to the sed expressions
which narrows the scope down to only "project" type lines.

Link: [1] https://github.com/OP-TEE/manifest/blob/3.19.0/fvp-ts.xml#L9
Link: [2] https://www.gnu.org/software/sed/manual/sed.html#sed-addresses
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Signed-off-by: Balint Dobszay <balint.dobszay@arm.com>