common: update buildroot to 2024.05
With more recent versions of GCC (v14.x), the build fails with implicit
declaration of function open_temp_exec_file(...) in libffi. The issue is
seen with GCC v14.1.1 and libffi v3.4.4. A fix for this issue was merged
in libffi v3.4.6, which is the version we get when updating Buildroot to
v2024.05.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>
Acked-by: Jerome Forissier <jerome.forissier@linaro.org>
Tested-by: Jerome Forissier <jerome.forissier@linaro.org> (vexpress-qemu_armv8a)
Tested-by: Joakim Bech <joakim.bech@linaro.org> (GCC 14.1.1/Fedora 40
           and GCC 11.4.0/Ubuntu 22.04)