Clone U-Boot from GitHub
Obtain U-Boot sources from GitHub [1] rather than from DENX [2]. [1] is
officially a mirror of [2] as mentioned in [3]. GitHub has ALWAYS
proved more reliable than any other source as far as Git repos are
concerned, for this OP-TEE build environment at least.

For the record, as of today, [2] returns an error page:
"""
HTTP 502: Waiting for GitLab to boot

It can take up to a few minutes for GitLab to boot completely.
"""

The error was reported to the U-Boot ML in [4].

[1] https://github.com/u-boot/u-boot
[2] https://source.denx.de/u-boot/u-boot
[3] https://lists.denx.de/pipermail/u-boot/2024-December/573566.html
[4] https://lists.denx.de/pipermail/u-boot/2025-July/594671.html

Change-Id: I663d94338f41d2239cf57ac073cfb5759d7a9b7f
Signed-off-by: Jerome Forissier <jerome.forissier@linaro.org>
Acked-by: Jens Wiklander <jens.wiklander@linaro.org>